requirements.txt
requests
python-dotenv
pandas
schedule